refactor: move slug regex
