+++ packages/outstatic/src/utils/schemas/helpers/slug-regex.test.ts
@@ -1,4 +1,4 @@
-import { slugRegex } from './slugRegex'
+import { slugRegex } from './slug-regex'
 
 describe('Regex Test', () => {
   it('should allow lowercase letters, numbers, and hyphens', () => {
