+++ packages/outstatic/src/utils/schemas/helpers/slug-regex.ts
